pyproject.toml
[tool.poetry]
name = "final-task-course-3"
version = "0.1.0"
description = ""
authors = ["Alexandr-lab-del <gordon675843@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.32.3"
pandas = "^2.2.2"
openpyxl = "^3.1.5"
python-dotenv = "^1.0.1"
pytest = "^8.3.2"


[tool.poetry.group.dev.dependencies]
pytest-cov = "^5.0.0"
flake8 = "^7.1.0"


[tool.poetry.group.lint.dependencies]
mypy = "^1.11.1"
black = "^24.4.2"
isort = "^5.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 119
exclude = '''
(
  /(
      \.eggs
    | \.git
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | dist
  )/
  | foo.py
)
'''

[tool.isort]
line_length = 119

[tool.mypy]
disallow_untyped_defs = true
no_implicit_optional = true
warn_return_any = true
exclude = 'venv'
